Breast_Intelligent_Recognition_Device/requirements.txt
albumentations==1.4.18
chardet==4.0.0
ml_collections==0.1.1
numpy==2.1.2
Pillow==10.4.0
scikit_learn==1.5.2
scipy==1.14.1
skimage==0.0
torch==2.3.0
torchvision==0.18.0
transformers==4.45.2
